split code + add youtube.js
